Add safeguard to prevent scanning if screen is off
Add an extra check to abort scanning if the screen is off. Usually this
should be caught by checking if onStart() has been called or if the app
is visible, but there are still reports of scanning happening even when
the screen is off. Adding this extra safeguard should remove these cases
entirely.

Flag: EXEMPT minor safeguard change

Bug: 386772617
Test: atest WifiTrackerLibTests
Change-Id: I12c812d91f90b7b00dc2168f04ce70ecca148525